// src/api/generated/model/internalCenterHttpHandlerPermissionItem.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */

export interface InternalCenterHttpHandlerPermissionItem {
  code?: string;
  description?: string;
  is_builtin?: boolean;
  name?: string;
  status?: string;
}