// src/api/generated/model/internalCenterHttpHandlerPermissionMutationResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalCenterHttpHandlerPermissionItem } from './internalCenterHttpHandlerPermissionItem';

export interface InternalCenterHttpHandlerPermissionMutationResponse {
  code?: number;
  data?: InternalCenterHttpHandlerPermissionItem;
  message?: string;
  trace_id?: string;
}