// src/api/generated/model/internalHandlerCommonActionResponse.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * zvas API
 * zvas 初始化阶段 API 文档。
 * OpenAPI spec version: 1.0
 */
import type { InternalHandlerActionData } from './internalHandlerActionData';

export interface InternalHandlerCommonActionResponse {
  code?: number;
  data?: InternalHandlerActionData;
  message?: string;
  trace_id?: string;
}